+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.5] - 2020-11-03
+### Added
+
+- Menux method slotMachine() (slots button)
+- Added new event onSlotMachine
+- Dice method isSlotMachine (check dice type)
+- Dice method getSlots
+
 ## [1.4.0] - 2020-04-15
 ### Added